hack together bloop compile (with infinite loop fixes on the bloop end)

+++ contrib/bloop/src/python/pants/contrib/bloop/tasks/compile/bloop_compile.py
@@ -6,6 +6,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from pants.backend.jvm.subsystems.zinc import Zinc
 from pants.backend.jvm.tasks.nailgun_task import NailgunTask
+from pants.base.exceptions import TaskError
 from pants.base.workunit import WorkUnitLabel
 from pants.java.jar.jar_dependency import JarDependency
 from pants.util.process_handler import subprocess
@@ -39,33 +40,41 @@ class BloopCompile(NailgunTask):
   _confs = Zinc.DEFAULT_CONFS
 
   def execute(self):
-    bsp_launcher_process = self.runjava(
+    runjava_args = dict(
       classpath=self.tool_classpath('bloop-compile-wrapper'),
       main='pants.contrib.bloop.compile.PantsCompileMain',
       jvm_options=[],
       # TODO: jvm options need to be prefixed with -J and passed to the LaumcherMain if we want to
       # use them!
       args=[
-        self.get_options().level,
+        'debug',
+        # self.get_options().level,
         '--',
       ] + [
         t.id for t in self.context.target_roots
       ],
       workunit_name='bloop-compile',
-      workunit_labels=[WorkUnitLabel.COMPILER],
-      do_async=True,
-      stdin=subprocess.PIPE,
-      stdout=subprocess.PIPE)
+      workunit_labels=[WorkUnitLabel.COMPILER])
 
-    msg = ''
-    while not msg:
-      msg = bsp_launcher_process.stdout.readline()
-    # assert msg == 'compile complete!'
-    self.context.log.info('msg: {}'.format(msg))
-
-    bsp_launcher_process.stdin.close()
-    bsp_launcher_process.stdout.close()
-    bsp_launcher_process.kill()
+    use_async = False
+    if use_async:
+      runjava_args.update(dict(
+        do_async=True,
+        stdin=subprocess.PIPE,
+        stdout=subprocess.PIPE))
+      bsp_launcher_process = self.runjava(**runjava_args)
+      msg = ''
+      while not msg:
+        msg = bsp_launcher_process.stdout.readline()
+      # assert msg == 'compile complete!'
+      self.context.log.info('msg: {}'.format(msg))
+      bsp_launcher_process.stdin.close()
+      bsp_launcher_process.stdout.close()
+      bsp_launcher_process.kill()
+    else:
+      rc = self.runjava(**runjava_args)
+      if rc != 0:
+        raise TaskError(exit_code=rc)
 
     for target in self.context.targets():
       classes_dir = self.context.products.get_data('bloop_classes_dir').get(target, None)
